wip: update the existing tests
